refactor: change the register and sidebar to temporary filenames
