+++ client/src/pages/EmployeeExpenses/EmployeeExpensesList.tsx
@@ -5,7 +5,7 @@ import { useAppSelector } from '../../actions/hooks'
 import { RootState } from '../../app/store'
 import { useDispatch } from 'react-redux'
 import { fetchPersonnel } from '../../reducers/personnel/personnelExpensesSlice'
-import Sidebar from '../../components/Sidebar/Sidebar'
+import Sidebar from '../../components/temp_SideBar/Sidebar'
 import { useLocation, useNavigate } from 'react-router-dom'
 import { useLanguage } from '../../contexts/LanguageContext'
 import { translate } from '../../utils/translationUtil'
