+++ client/src/pages/temp_dashboard/temp_dashboard.tsx
@@ -9,7 +9,7 @@ import { RootState } from '../../app/store'
 import { TablePlanningB } from '../../components/Table/TablePlanningB.component'
 import { fetchAllClientData } from '../../reducers/table/tableSlice'
 import { fetchGraphData } from '../../reducers/graph/graphSlice'
-import Sidebar from '../../components/Sidebar/Sidebar'
+import Sidebar from '../../components/temp_SideBar/Sidebar'
 import Btn from '../../components/Button/Button'
 import { useLocation, useNavigate } from 'react-router-dom'
 import TablePlanningA from '../../components/Table/TablePlanningA'
