+++ client/src/pages/temp_register/temp_register.tsx
